+++ pysui_gadgets/package/cmds/structs.py
@@ -86,12 +86,15 @@ def print_module_structs(modules: dict[str, SuiMoveModule], args: Namespace) -> 
         for struct_name, struct_def in mod_def.structs.items():
             sig = _struct_parm_abilities(struct_def.type_parameters, "")
             sig += _struct_parm_types(struct_def.abilities.abilities, "")
-            print(f"    {struct_name} {sig} {{")
-
+            print(f"\nStruct: {struct_name} {sig} {{")
+            max_field_name = 0
             for field in struct_def.fields:
-                print("", end="")
-                print(field.to_json(indent=2))
-                sig = f"        {field.name:30s}: "
-                sig = _field_signature(field, sig)
-                # print(sig)
+                if len(field.name) > max_field_name:
+                    max_field_name = len(field.name)
+            max_field_name += 1
+            for field in struct_def.fields:
+                if not args.short_form:
+                    print(field.to_json(indent=2))
+                else:
+                    print(f"    {field.name:{max_field_name}s}: {_field_signature(field,'')}")
             print("}")